Open the editor and android studio based on the FreeDesktop entries
This can make good use of more testing. But now it can probably do all the apple
version did. Even if it's not as clean. There is a lot of parsing of the FreeDesktop
files, and the code for finding the application by name can be very slow, although
it didn't take long at all on the tests I did.