Add open xcode with xed command
# What is `xed`

`xed` is the way to launch using Xcode, which the system recognizes as the default. (Able to use `xcode-select -p` to see which Xcode is recognized as default).

Starting with the name `Xcode` may fail. For example, if you have multiple Xcode versions installed, they are named `Xcode-13.x`, `Xcode-12.x`, and so on. Therefore `open -a Xcode` will fail.

xed is a kind of wrapper for the open command that opens Xcode as recognized by the system. This is very robust and useful, because it allows developers to find whatever development environment they are in.